requirements.txt
aiohttp~=3.10.10
discord~=2.3.2
python-dotenv~=1.0.1